General clean up and fixes.

+++ src/1.foundation/2.game_loop_and_sdl/2.0.game_loop_and_sdl/src/fps_counter.cpp
@@ -2,7 +2,7 @@
 
 #include "./re/utils/logger.h"
 
-FPSCounter::FPSCounter() : timer(new Timer(1000, true)) {
+FPSCounter::FPSCounter() : timer(new Timer((Uint32) 1000, true)) {
     timer->Start();
 }
